+++ lib/src/view/DataModelWindow.cxx
@@ -29,16 +29,12 @@
 
 #include <QHBoxLayout>
 #include <QHeaderView>
-#include <QPushButton>
 #include <QToolButton>
 #include <QGroupBox>
 #include <QFileDialog>
 #include <QMessageBox>
-#include <QSortFilterProxyModel>
 #include <QScrollBar>
-#include <QApplication>
 #include <QScrollArea>
-#include <QMenu>
 
 using namespace OT;
 
@@ -410,58 +406,5 @@ void DataModelWindow::updateTableView()
 
   // update sample size label
   sampleSizeLabel_->setText(QString::number(dataModel_->getSample().getSize()));
-}
-
-EditableExportableTableView::EditableExportableTableView(QWidget* parent)
-  : ExportableTableView(parent)
-  , addRowAction_(0)
-  , removeRowAction_(0)
-{
-  addRowAction_ = new QAction(QIcon(":/images/list-add.png"), tr("Add row"), this);
-  removeRowAction_ = new QAction(QIcon(":/images/list-remove.png"), tr("Remove row(s)"), this);
-
-  connect(addRowAction_, SIGNAL(triggered()), this, SLOT(addRow()));
-  connect(removeRowAction_, SIGNAL(triggered()), this, SLOT(removeRows()));
-  setContextMenuPolicy(Qt::CustomContextMenu);
-}
-
-// show the context menu when right clicking
-void EditableExportableTableView::contextMenu(const QPoint & pos)
-{
-  QMenu * contextMenu(new QMenu(this));
-  contextMenu->addAction(exportAction_);
-  contextMenu->addAction(addRowAction_);
-  contextMenu->addAction(removeRowAction_);
-  if (exportableAsImageAction_)
-    contextMenu->addAction(exportableAsImageAction_);
-  contextMenu->popup(this->mapToGlobal(pos));
-}
-
-void EditableExportableTableView::addRow()
-{
-  QAbstractItemModel * sourceModel = dynamic_cast<QSortFilterProxyModel*>(model())->sourceModel();
-  sourceModel->insertRows(sourceModel->rowCount()-1, 1);
-
-}
-
-void EditableExportableTableView::removeRows()
-{
-  QSortFilterProxyModel* myProxy = dynamic_cast<QSortFilterProxyModel*>(model());
-  QAbstractItemModel * sourceModel = myProxy->sourceModel();
-  QItemSelection selection(selectionModel()->selection());
-  QList<int> rows;
-  foreach( const QModelIndex & index, myProxy->mapSelectionToSource(selection).indexes() ) {
-    rows.append( index.row() );
-  }
-  std::sort( rows.begin(), rows.end() );
-  int prev = -1;
-  for( int i = rows.count() - 1; i >= 0; i -= 1 ) {
-    int current = rows[i];
-    if( current != prev ) {
-      sourceModel->removeRows( current, 1 );
-      prev = current;
-    }
-  }
-
 }
 }
